refactor: now utils is an internal package
Let us make internal everything because this is not a library.
